Solves the ISSUE #446. No need to update the CURRENT_ENGINE_RELEASE property when it has value.
In current code when the CURRENT_ENGINE_RELEASE contains the same version as in remote web, it overwrites
the property with the empty string, but in OracleDB there is no difference between empty string and null,
and it stores a null value, so when the plugin read all properties and it tries to store them in a HashMap,
the plugin fails with a NullPointerException.

And the code executes one of twice for the same version, when there is no need to save it until it changes.