Change the GrokAssembly binary with one that works with dotnet Core.
GrokAssembly was loading the binaries to examine in order to identify it using Reflection.

The probleme was when the examined assembly was dependant of some .NET Core specific API, the assembly would refuse to load.

This new version of GrokAssembly lookup the information from the assembly in a way that does not fail when the Assembly is missing some API's.

You can check the source code of this binary here : https://github.com/thomasd-/GrokAssembly