Issue-1751: Make Maven plugin skip artifact resolution for artifacts that are already resolved.
The artifact resolution code flattens out all dependencies, and tries to resolve them from the repositories defined in the project the plugin is being executed for. This is wrong for transitive dependencies. For example, given dependency chain A -> B -> C, C may be in a repository only mentioned in B's POM. Running the plugin for A would fail in this case. Since the mojos making use of the artifact resolution code are already configured to require artifact resolution before running, it is better to just use the artifacts Maven has resolved. The only artifacts that need to be manually resolved are reactor projects, and maybe dependencyManagement dependencies that are not actually depended upon.